Fix crash since ldlib updated to 1.0.28.a./修复ldlib升级到1.0.28.a导致的崩溃问题。
Fix item in share inventory not drop when hugeitembus was removed./修复巨型输入总线在破坏时催化剂栏中物品无法掉落的问题。
Change creative input hatch's fluid slot's max amount to long.max_value./修改创造模式输入仓的流体数量为Long的上限值。
Add recipe for advanced terminal./添加高级终端，可以指定线圈等级，无仓室模式，重复结构次数。

+++ src/main/java/com/hepdd/gtmthings/common/block/machine/multiblock/part/CreativeInputHatchPartMachine.java
@@ -61,7 +61,7 @@ public class CreativeInputHatchPartMachine extends TieredIOPartMachine implement
     }
 
     protected NotifiableFluidTank createTank() {
-        return new NotifiableFluidTank(this,SLOT_COUNT,Integer.MAX_VALUE,IO.IN);
+        return new NotifiableFluidTank(this,SLOT_COUNT,Long.MAX_VALUE,IO.IN);
     }
 
     @Override
@@ -96,7 +96,7 @@ public class CreativeInputHatchPartMachine extends TieredIOPartMachine implement
             for (int i = 0; i < SLOT_COUNT; i++) {
                 if (fluidMap.containsKey(i)) {
                     var mFluid = this.creativeTanks[i].getFluid().copy();
-                    mFluid.setAmount(Integer.MAX_VALUE);
+                    mFluid.setAmount(Long.MAX_VALUE);
                     this.tank.setFluidInTank(i,mFluid);
                 } else {
                     if (!this.tank.getFluidInTank(i).isEmpty()) {
